Created automated NLP Web Scraping Script
This is a contribution to start a project with you. It contains a modified version of your python script that allows you to scrape a webpage and perform automated sentiment analysis. Instead of parsing an offline webpage, this script asks you programmatically for a link and automatically downloads and parses it. The end product is a printed out sentiment analysis and a piechart that automatically breaks-down positive and negative language graphically.

The hope is to work with you in order to further develop this project into a Flask WebApp. The goal is to give users the ability to perform sophisticated sentiment analysis of any webpage easily and conveniently.